Cleanup spacing and namespaces
