+++ Cql/Cql.Packaging/PackagerService.cs
@@ -1,9 +1,8 @@
-﻿using Hl7.Cql.Packaging;
-using Hl7.Cql.Packaging.ResourceWriters;
+﻿using Hl7.Cql.Packaging.ResourceWriters;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Hl7.Cql.Packager;
+namespace Hl7.Cql.Packaging;
 
 internal class PackagerService
 {
